requirements.txt
fastapi
uvicorn
streamlit
python-dotenv
langchain-groq
langchain-openai
langchain-tavily
langgraph
langchain-core
langchain-community
langchain-cohere
pydantic
requests
cohere
faiss-cpu
pdfplumber
PyPDF2
chromadb
python-multipart
